Add files via upload
A large number of spam mails fall into our mailbox every day. The damage of spam mails greatly affects personal accounts and companies. I developed this project to affect all this damage. While developing this project, I used all mails and spam mails list. My method of separating spam mails was on the basis of "words." 10 most used words in spam mails. The programming language I developed is Python.Train and the results of the test data are also included in the project.